requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
jinja2==3.1.4
python-multipart==0.0.9
apscheduler==3.10.4
openai==1.30.1
python-dotenv==1.0.1
